refactor!(feeder): Use more descriptive name to match design pattern used
