+++ fang-mcb-project/src/control/turret/feeder/simple_feeder_subsystem/m2006_simple_feeder_subsystem_factory.cpp
@@ -1,7 +1,7 @@
-#include "m2006_simple_feeder_subsystem_maker.hpp"
+#include "m2006_simple_feeder_subsystem_factory.hpp"
 namespace fang::turret
 {
-    M2006SimpleFeederSubsystemMaker::SimpleFeederSubsystemPtr M2006SimpleFeederSubsystemMaker::Make(Drivers& drivers, const Config& config)
+    M2006SimpleFeederSubsystemFactory::SimpleFeederSubsystemPtr M2006SimpleFeederSubsystemFactory::Make(Drivers& drivers, const Config& config)
     {
         auto feederSystem{M2006SimpleFeederFactory::Make(drivers, {config.feedMotorConfig, config.feederSystemConfig})};
         return std::make_unique<SimpleFeederSubsystem>(drivers, std::move(feederSystem), config.feederSubsystemConfig);
