+++ fang-mcb-project/src/control/turret/feeder/simple_feeder_subsystem/m2006_simple_feeder_subsystem_factory.hpp
@@ -3,7 +3,7 @@
 #include "simple_feeder_subsystem.hpp"
 #include "driver/drivers.hpp"
 #include "control/turret/feeder/simple_feeder_system/simple_feeder_system.hpp"
-#include "control/turret/feeder/simple_feeder_system/m2006_simple_feeder_maker.hpp"
+#include "control/turret/feeder/simple_feeder_system/m2006_simple_feeder_factory.hpp"
 
 #include "trap/motor/dji_m2006.hpp"
 
@@ -12,7 +12,7 @@ namespace fang::turret
     /**
      * Gives a basic feeder that utilizes an M2006 motor
      */
-    class M2006SimpleFeederSubsystemMaker
+    class M2006SimpleFeederSubsystemFactory
     {
     public:
         using FeedMotor = trap::motor::DjiM2006;
@@ -24,7 +24,7 @@ namespace fang::turret
             SimpleFeederSystem::Config feederSystemConfig;
         };
 
-        M2006SimpleFeederSubsystemMaker(Drivers& drivers, const Config& config);
+        M2006SimpleFeederSubsystemFactory(Drivers& drivers, const Config& config);
         static SimpleFeederSubsystemPtr Make(Drivers& drivers, const Config& config);
 
     };
