+++ fang-mcb-project/src/robot/variant/pierce/pierce.cpp
@@ -14,7 +14,7 @@ namespace control
             mk_commandConfig{config.commandConfig},
             kMappingConfig_{config.mappingConfig},
             m_imu{drivers.bmi088},
-            feeder_{std::move(fang::turret::M2006SimpleFeederSubsystemMaker::Make(drivers, config.subsystemConfig.feederConfig))},
+            feeder_{std::move(fang::turret::M2006SimpleFeederSubsystemFactory::Make(drivers, config.subsystemConfig.feederConfig))},
             m_chassis{drivers, mk_subsystemConfig.chassisConfig},
             m_gimbal{drivers, m_imu, mk_subsystemConfig.gimbalConfig},
             m_booster{drivers, mk_subsystemConfig.boosterConfig},
